[백엔드] 게임 승패 기록

+++ backend/src/main/java/com/chipchippoker/backend/api/gameresult/service/GameResultServiceImpl.java
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 public class GameResultServiceImpl implements GameResultService {
 	private final GameResultRepository gameResultRepository;
 
-	public void saveGameResult(List<String> memberList, List<Integer> resultCoinList, Integer gameRoomId,
+	public void saveGameResult(List<String> memberList, List<Integer> resultCoinList, Long gameRoomId,
 		String gameMode) {
 		GameResult gameResult = GameResult.newGameResult(memberList, resultCoinList, gameRoomId, gameMode);
 		gameResultRepository.save(gameResult);
